Fix #3967 - Remove maven-jsf-plugin
- Remove unnecessary code (cast, redundant modifiers)

+++ src/main/java/org/primefaces/component/treetable/TreeTable.java
@@ -487,7 +487,7 @@ public class TreeTable extends TreeTableBase {
             if (rows > 0) {
                 int first = getFirst();
 
-                return (int) (first / rows);
+                return first / rows;
             }
             else {
                 return 0;
